Move PlatformLive.ExecutorUtil to Executor companion object. private zio.internal.Sync
