+++ core/js/src/main/scala/zio/internal/DefaultExecutors.scala
@@ -1,0 +1,3 @@
+package zio.internal
+
+private[internal] abstract class DefaultExecutors
