+++ core/js/src/main/scala/zio/internal/Sync.scala
@@ -1,6 +1,6 @@
 package zio.internal
 
-object Sync {
+private[zio] object Sync {
   def apply[A](anyRef: AnyRef)(f: => A): A = {
     val _ = anyRef
     f
