+++ core/jvm/src/main/scala/zio/internal/PlatformLive.scala
@@ -16,7 +16,6 @@
 
 package zio.internal
 
-import java.util.concurrent.{ Executor => _, _ }
 import java.util.{ Collections, WeakHashMap, Map => JMap }
 import zio.Cause
 import zio.internal.stacktracer.Tracer
@@ -40,7 +39,8 @@ object PlatformLive {
    * */
   lazy val Benchmark = makeDefault(Int.MaxValue).withReportFailure(_ => ()).withTracing(Tracing.disabled)
 
-  final def makeDefault(yieldOpCount: Int = 2048): Platform = fromExecutor(ExecutorUtil.makeDefault(yieldOpCount))
+  final def makeDefault(yieldOpCount: Int = defaultYieldOpCount): Platform =
+    fromExecutor(Executor.makeDefault(yieldOpCount))
 
   final def fromExecutor(executor0: Executor) =
     new Platform {
@@ -69,70 +69,7 @@ object PlatformLive {
     }
 
   final def fromExecutionContext(ec: ExecutionContext): Platform =
-    fromExecutor(Executor.fromExecutionContext(1024)(ec))
+    fromExecutor(Executor.fromExecutionContext(defaultYieldOpCount)(ec))
 
-  object ExecutorUtil {
-    final def makeDefault(yieldOpCount: Int): Executor =
-      fromThreadPoolExecutor(_ => yieldOpCount) {
-        val corePoolSize  = Runtime.getRuntime.availableProcessors() * 2
-        val maxPoolSize   = corePoolSize
-        val keepAliveTime = 1000L
-        val timeUnit      = TimeUnit.MILLISECONDS
-        val workQueue     = new LinkedBlockingQueue[Runnable]()
-        val threadFactory = new NamedThreadFactory("zio-default-async", true)
-
-        val threadPool = new ThreadPoolExecutor(
-          corePoolSize,
-          maxPoolSize,
-          keepAliveTime,
-          timeUnit,
-          workQueue,
-          threadFactory
-        )
-        threadPool.allowCoreThreadTimeOut(true)
-
-        threadPool
-      }
-
-    final def fromThreadPoolExecutor(yieldOpCount0: ExecutionMetrics => Int)(
-      es: ThreadPoolExecutor
-    ): Executor =
-      new Executor {
-        private[this] def metrics0 = new ExecutionMetrics {
-          def concurrency: Int = es.getMaximumPoolSize()
-
-          def capacity: Int = {
-            val queue = es.getQueue()
-
-            val remaining = queue.remainingCapacity()
-
-            if (remaining == Int.MaxValue) remaining
-            else remaining + queue.size
-          }
-
-          def size: Int = es.getQueue().size
-
-          def workersCount: Int = es.getPoolSize()
-
-          def enqueuedCount: Long = es.getTaskCount()
-
-          def dequeuedCount: Long = enqueuedCount - size.toLong
-        }
-
-        def metrics = Some(metrics0)
-
-        def yieldOpCount = yieldOpCount0(metrics0)
-
-        def submit(runnable: Runnable): Boolean =
-          try {
-            es.execute(runnable)
-
-            true
-          } catch {
-            case _: RejectedExecutionException => false
-          }
-
-        def here = false
-      }
-  }
+  final val defaultYieldOpCount = 2048
 }
