+++ core/jvm/src/main/scala/zio/internal/Sync.scala
@@ -1,5 +1,5 @@
 package zio.internal
 
-object Sync {
+private[zio] object Sync {
   def apply[A](anyRef: AnyRef)(f: => A): A = anyRef.synchronized { f }
 }
